requirements.txt
# AI will API - Python Dependencies
# FastAPI and ASGI
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
python-multipart>=0.0.6

# Pydantic (data validation)
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
email-validator>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
sqlalchemy>=2.0.0,<3.0.0
asyncpg>=0.29.0
aiosqlite>=0.19.0
alembic>=1.13.0

# Redis (caching, rate limiting, idempotency)
redis>=5.0.0
aioredis>=2.0.0

# HTTP Client (for external APIs)
httpx>=0.26.0

# SSE (Server-Sent Events)
sse-starlette>=1.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0  # for TestClient

# Development
black>=24.1.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0

# Type stubs
types-passlib>=1.7.0
types-python-jose>=3.3.0
